### 2022-07-19 #4592 @editorialbot
[REVIEW]: SimSGamE : Scheduling simulator for modern game engines
**Submitting author:** @baptisteCoye (<a href="http://orcid.org/0000-0002-7619-4492">Baptiste Coye</a>)
**Repository:** https://github.com/baptisteCoye/SimSGame
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
