### 2022-07-19 #4592 @danielskatz
👋 @bcoye - In your paper.md file, you have an endash rather then a hyphen in the ORCID for Raymond Namyst.  Please change this, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
