### 2022-07-26 #4592 @danielskatz
I'm on vacation, and am not sure I saw your email - can you resend it?

Diving into the error message above, it still gets to

Problem with ORCID (0000−0001−7734−1258) for Raymond Namyst. ORCID looks malformed (Theoj::Error)

And it still looks to me like all of the dashes are endashes `−` rather than hyphens `-`.

Try changing it to `0000-0001-7734-1258`
