### 2022-07-26 #4592 @hwloidl
@bcoye Overall `paper.md` with Summary and Statement of NEed are good; however I suggest to add a 3rd para in Summary which is more specific about the problem you want to solve (scheduling of a task graph) and the approach you are taking; a short version of paras 3 and 4 from your EuroPar22 paper would work well to this end
