### 2022-07-26 #4592 @hwloidl
@bcoye regarding Dan's comment above on "gitlab instance that is limited to INRIA " can you confirm that this repo is open and available without an account? I think I managed to clone the repo without loging in, so this part seems ok to me at this point.
