### 2022-07-27 #4592 @danielskatz
> @bcoye regarding Dan's comment above on "gitlab instance that is limited to INRIA " can you confirm that this repo is open and available without an account? I think I managed to clone the repo without loging in, so this part seems ok to me at this point.

This is not sufficient. The community must have a way to create issues as well as being to create PRs, not just clone the repo. The latter is open source, but the former is needed for this to be a community project and meet the JOSS community guidelines.
