### 2022-07-27 #4592 @danielskatz
> @danielskatz can you double check that these messages are reaching the author? I note that the handle (@bcoye) is not bold-faced and not clickable; should he be added to the Review repo to engage in discussions?

Yes, there's a problem here - thanks for catching it.  I'll try to figure it out
