### 2022-07-27 #4592 @baptisteCoye
Hi, sorry couldn't get the notifications.
Yes you are right, i thought a classic gitlab account would be ok to create issues on a public git but it seems it doesn't work.
I'll recreate the repository on github, and make the necessary changes regarding the ORCID and add the paragraph @hwloidl was refering to earlier.
Should I send you the new git repository here?
