+++ paper.md
@@ -1,0 +1,113 @@
+---
+#title SimSGamE : Scheduling simulator for modern game engines
+
+tags:
+  - Task Programming
+  - Video-game engine
+  - Scheduling
+  - Soft Real Time Scheduling
+  - C++
+
+authors:
+  - name: Mustapha Regragui
+    orcid: 0000-0001-6316-8079
+    equal-contrib: true
+    affiliation: "1"
+  - name: Baptiste Coye
+    orcid: 0000-0002-7619-4492
+    equal-contrib: true 
+    affiliation: "1, 2"
+  - name: Laercio Lima Pilla
+    orcid: 0000-0003-0997-586X
+    equal-contrib: true 
+    affiliation: "1, 3"
+  - name: Raymond Namyst
+    orcid: 0000−0001−7734−1258
+    equal-contrib: true
+    affiliation: "1, 4, 5"
+  - name: Denis Barthou
+    orcid: 0000-0001-7110-2037
+    equal-contrib: true
+    affiliation: "1, 5, 6"
+
+affiliations:
+ - name: Inria Bordeaux, France
+   index: 1
+ - name: Ubisoft Bordeaux, France
+   index: 2
+ - name: CNRS, France
+   index: 3
+ - name: Univ. Bordeaux, France
+   index: 4
+ - name: Labri, France
+   index: 5
+ - name: Bordeaux INP, France
+   index: 6
+   
+date: 27 June 2022
+bibliography: paper.bib
+---
+
+# Summary
+
+The video game market is valued at over 100$ billion USD and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
+The variety of devices running games is very large, from personnal computers, consoles and smartphones to Cloud gaming servers.
+In order to develop these games more easily and allow to port them on different platforms, developpers use a key software component called the game engine.
+
+Game engines are at the heart of the design of modern video games. They handle most of the components of the game, and allows the generation of each frame(image) displayed on the player screen. They aim to keep a high frame (30 to 60 Frames per seconds) rate by scheduling the tasks required to complete such computations.
+These tasks are organized in a soft real-time, parallel task graph, which is a scenario very few works have focused on, or adapted scheduling algorithms to.
+
+
+# Statement of need
+
+SimSGamE was created in order to study the scheduling of game engines. Given the lack of studies on this problem, our efforts have been dedicated to finding and adapting algorithms and heuristics proposed in other contexts. We find that there is value in bringing to light new applications and knowledge on existing algorithms.
+
+The software simulates the behavior of game engines using two files : 
+
+- a dependency file describing links between tasks and the structure of the game engine to be simulated
+- a tasks file describing key values for tasks (steps, mean, minimum, maximum, std_dev) for both moderated and intensive use of the engine.
+
+Using these two files, SimSGamE generates randomized tasks composing a frame respecting the statistical repartition and their dependencies (generation may be seeded in order to enable direct comparisons). The simulation also mimic different load evolution (from moderate use 0% to intensive 100%) and evaluates multiple scheduling techniques to estimate frame span and metrics giving informations regarding heuristics efficiencies.
+
[…]
