+++ paper.bib
@@ -1,0 +1,250 @@
+@article{trietsch2012modeling,
+  title={Modeling activity times by the Parkinson distribution with a lognormal core: Theory and validation},
+  author={Trietsch, Dan and Mazmanyan, Lilit and Gevorgyan, Lilit and Baker, Kenneth R},
+  journal={European Journal of Operational Research},
+  volume={216},
+  number={2},
+  pages={386--396},
+  year={2012},
+  publisher={Elsevier},
+  doi={10.1016/j.ejor.2011.07.054}
+}
+
+@INPROCEEDINGS{nascimento2021effectively,
+  author={Nascimento, Flávia Maristela S. and Lima, George},
+  booktitle={2021 IEEE 27th Real-Time and Embedded Technology and Applications Symposium (RTAS)}, 
+  title={Effectively Scheduling Hard and Soft Real-Time Tasks on Multiprocessors}, 
+  year={2021},
+  volume={},
+  number={},
+  pages={210-222},
+  doi={10.1109/RTAS52030.2021.00025}
+}
+
+@INPROCEEDINGS{dai2021timing,
+  author={Dai, Gaoyang and Mohaqeqi, Morteza and Yi, Wang},
+  booktitle={2021 IEEE 27th International Conference on Embedded and Real-Time Computing Systems and Applications (RTCSA)}, 
+  title={Timing-Anomaly Free Dynamic Scheduling of Periodic DAG Tasks with Non-Preemptive Nodes}, 
+  year={2021},
+  volume={},
+  number={},
+  pages={119-128},
+  doi={10.1109/RTCSA52859.2021.00022}
+}
+
+@misc{gds2019,
+author={Bret Alfieri},
+howpublished={\url{https://github.com/GameTechDev/GTS-GamesTaskScheduler}},
+title={Intel Games Task Scheduler},
+year=2019,
+month=Jan,
+note={Accessed: 2022-01-04}
+}
+
+@article{acun2016power,
+  title={Power, reliability, and performance: One system to rule them all},
+  author={Acun, Bilge and Langer, Akhil and Meneses, Esteban and Menon, Harshitha and Sarood, Osman and Totoni, Ehsan and Kal{\'e}, Laxmikant V},
+  journal={Computer},
+  volume={49},
+  number={10},
+  pages={30--37},
+  year={2016},
+  publisher={IEEE},
+  doi={10.1109/MC.2016.310}
+}
+
+@book{gregory2018game,
+  author = {Gregory, Jason},
+  edition = 3,
+  keywords = {Engineering ITeG_233 architecture game programming},
+  publisher = {Taylor and Francis Ltd.},
+  title = {Game engine architecture},
+  year = 2018
+}
+
+@misc{gaming,
+title={{Global Gaming Market - Growth, Trends, Covid-19 Impact, and Forecasts (2022--2027)}},
+howpublished={\url{https://www.mordorintelligence.com/industry-reports/global-gaming-market}},
+note={Accessed: 2022-01-26},
+author={Mordor~Intelligence}
+}
+
+@misc{unity,
+title={{Unity User Manual 2020.3 (LTS)}},
+howpublished={\url{https://docs.unity3d.com/Manual/UnityManual.html}},
+note={Accessed: 2022-01-26},
+key={Unity Technologies}
+}
+
+@misc{unreal,
+title={{Unreal Engine 4 Documentation}},
+howpublished={\url{https://docs.unrealengine.com/4.27/en-US/}},
+note={Accessed: 2022-01-26},
+key={Unreal}
+}
+
+@book{leung2004handbook,
+  title={Handbook of scheduling: algorithms, models, and performance analysis},
+  author={Leung, Joseph YT},
+  year={2004},
[…]
