### 2022-07-27 #4592 @hwloidl
@baptisteCoye while you are at making the changes, this item in my reviewer checklist is currently not covered in the paper: Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
I suggest you you add are short section on this and add the information mentioned above, pointing to the github repo
