### 2022-07-27 #4592 @hwloidl
@baptisteCoye Another item still to cover in paper.md is: "State of the field: Do the authors describe how this software compares to other commonly-used packages?"
You just shortly mention "lack of studies". I think your Europar22 paper has a bit more on this in the intro. One para on this (probably in the intro, or in a new background section) should address this
