### 2022-07-27 05098c2
Update paper.md
Add paragraph describing the problem we adress in the code
