### 2022-07-27 #4592 @baptisteCoye
@danielskatz Here is the new git repository with the right access: (if new submission is needed let me know)
https://github.com/baptisteCoye/SimSGame

I'll add the community guidelines and state of the field right away.
