### 2022-07-28 #4592 @danielskatz
@baptisteCoye - The paper.md still has the same problem with incorrect hyphens for an ORCID - see https://github.com/openjournals/joss-reviews/issues/4592#issuecomment-1195546027

And yes, the community guidelines should be in the readme, and can also be in the paper but don't have to be there.
