### 2022-07-28 #4592 @danielskatz
actually, now I see that maybe the orcid was changed - I think https://github.com/baptisteCoye/SimSGame/pull/1 should fix it
