### 2022-07-28 #4592 @hwloidl
> > @danielskatz Here is the new git repository with the right access: (if new submission is needed let me know) https://github.com/baptisteCoye/SimSGame
> > I'll add the community guidelines and state of the field right away.
>
> @baptisteCoye how I understand this is that the community guidelines should be part of the readme.md in the repository and not in the paper.

@baptisteCoye As @azoitl says, community guidelines are better suited for the documentation in the readme.md, but they should cover the 3 areas mentioned in the comment (and checklist item) above
