### 2022-07-28 #4592 @baptisteCoye
Sorry I'm not sure I understood correctly, should I modify the readme.md and extend the Community guidelines?
