### 2022-07-28 #4592 @hwloidl
> Sorry I'm not sure I understood correctly, should I modify the readme.md and extend the Community guidelines?

@baptisteCoye Yes there should be a section at the end of the readme.md, covering: "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"
