### 2022-07-29 b4f9aa0
Update README.md
Expand Community Guidelines
