### 2022-08-01 #4592 @azoitl
The error message is:
`Error running filter /usr/local/share/openjournals/data/filters/normalize-metadata.lua:
PandocLuaError "Cannot get Attr from TypeNil"
stack traceback:
	...l/share/openjournals/data/filters/normalize-metadata.lua:95: in function 'Meta'
`
Any tips what this means?
