### 2022-08-01 #4592 @xuanxu
It looks like the error is caused by the paper's metadata [having no value](https://github.com/baptisteCoye/SimSGame/blob/main/paper.md?plain=1#L2) for `title`.
