### 2022-08-02 #4592 @danielskatz
@xuanxu - do you think we could create a more meaningful error message for future occurrences?
