### 2022-08-02 #4592 @danielskatz
@baptisteCoye - in your paper.md, the title is commented out. Can you remove the leading #?  Then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
