### 2022-08-02 054ef7f
Update paper.md
Remove # before title
