+++ paper.md
@@ -1,5 +1,5 @@
 ---
-#title SimSGamE : Scheduling simulator for modern game engines
+title: SimSGamE : Scheduling simulator for modern game engines
 
 tags:
   - Task Programming
