### 2022-08-02 8a14403
Update paper.md
Reformat paper.md for pdf compilation
