### 2022-08-02 a9d8793
Update paper.md
Fix typo issues and wrong generation of Title
