### 2022-08-02 #4592 @azoitl
@baptisteCoye your references are still not correct. Please see the [citation rules](https://pandoc.org/MANUAL.html#extension-citations) for more information.
