### 2022-08-02 d86ef20
Update paper.md
Update references
