+++ paper.md
@@ -46,7 +46,7 @@ bibliography: paper.bib
 
 # Summary
 
-The video game market is valued at over 100$ billion USD and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
+The video game market is valued at over 100$ billion USD [@gaming] and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
 The variety of devices running games is very large, from personnal computers, consoles and smartphones to Cloud gaming servers.
 In order to develop these games more easily and allow to port them on different platforms, developpers use a key software component called the game engine.
 
@@ -110,14 +110,12 @@ Metrics
 # State of the field
 
 Scheduling is a well known issue that have been widely studied however, given the lack of studies on this particular scheduling problem, our efforts have been dedicated to finding and adapting algorithms and heuristics proposed in other context.
-Video games work as soft real-time interactive simulations, however Real-Time scheduling often considers independent reccuring tasks where earliest deadline first(EDF) heuristics are employed. Nonetheless, the game engine contains dependent tasks with an entire task graph to be computed each frame and all tasks share the same due date, obstructing the use of EDF heuristics. Moreover, parallel task scheduling usually models tasks using multiple resources simultaneously, but the tasks follows a fork-join model internally. An algorithm called DynFed was proposed to schedule parallel tasks with dependencies in real time systems by Dai, Mohaqeqi and Yi but it focuses on periodic, independent tasks whose parallel subtasks have dependencies while our problem contains periodic tasks with dependencies whose parallel subtasks ar independant.
-The Game Task Scheduler developped by Intel reflects this tasks and subtasks organisation but is developped as a tool to be integrated in a game engine and not as a benchmark system allowing to test different heurisitcs and evaluating its performances.
+Video games work as soft real-time interactive simulations [@gregory2018game], however Real-Time scheduling often considers independent reccuring tasks where earliest deadline first (EDF) heuristics are employed [@nascimento2021effectively]. Nonetheless, the game engine contains dependent tasks with an entire task graph to be computed each frame and all tasks share the same due date, obstructing the use of EDF heuristics. Moreover, parallel task scheduling usually models tasks using multiple resources simultaneously, but the tasks follows a fork-join model internally. An algorithm called DynFed was proposed to schedule parallel tasks with dependencies in real time systems by Dai, Mohaqeqi and Yi [@dai2021timing] but it focuses on periodic, independent tasks whose parallel subtasks have dependencies while our problem contains periodic tasks with dependencies whose parallel subtasks ar independant.
+The Game Task Scheduler developped by Intel [@gds2019] reflects this tasks and subtasks organisation but is developped as a tool to be integrated in a game engine and not as a benchmark system allowing to test different heurisitcs and evaluating its performances.
 
 # Acknowledgements
 
 We acknowledge contributions from Cédric Dumondelle and Hervé Hubele during the genesis of this project.
 
 # References
-
-References can be found in paper.bib
 
