### 2022-08-03 #4592 @hwloidl
@baptisteCoye I had a look at the generated .pdf and have adjusted my review accordingly (above). The points remaining are, I think, a couple of minor issues and typos in the paper (below) and the issue about citations mentioned by @azoitl
