### 2022-08-03 #4592 @hwloidl
@baptisteCoye Here the list of typos I mention above:
Minor issues and typos in paper.md:

personnal computers -> personal computers

developpers -> developers

and allows -> and allow

[reorder] frame (30 to 60 Frames per seconds) rate -> frame rate (30 to 60 Frames per seconds)

[I assume you want a compound noun in this case] precedences constraints -> precedence constraints

execution’s correctness which -> the execution’s correctness, which

[not sure what you mean here:]  "leads their organisation"
maybe rephrase as "leads to a ... DAG" or "is organised as a ... DAG"

is focus on modeling -> is focused on modeling

after "answering the following three questions: -" make sure these come up as
items in an itemisation; at the moment these are inline text!

simulation also mimic -> simulation also mimics

informations -> information

it’s -> their

independent reccuring ->  independently reccuring

omputed each frame -> omputed at each frame

ar independant -> are independent

developped ->  developed

this tasks and subtasks -> these tasks and subtasks

its performances -> its performance
