### 2022-08-04 254c9cc
Update paper.md
Fix minor issues and typo
