+++ paper.md
@@ -47,15 +47,16 @@ bibliography: paper.bib
 # Summary
 
 The video game market is valued at over 100$ billion USD [@gaming] and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
-The variety of devices running games is very large, from personnal computers, consoles and smartphones to Cloud gaming servers.
-In order to develop these games more easily and allow to port them on different platforms, developpers use a key software component called the game engine.
+The variety of devices running games is very large, from personal computers, consoles and smartphones to Cloud gaming servers.
+In order to develop these games more easily and allow to port them on different platforms, developers use a key software component called the game engine.
 
-Game engines are at the heart of the design of modern video games. They handle most of the components of the game, and allows the generation of each frame(image) displayed on the player screen. They aim to keep a high frame (30 to 60 Frames per seconds) rate by scheduling the tasks required to complete such computations.
+Game engines are at the heart of the design of modern video games. They handle most of the components of the game, and allow the generation of each frame(image) displayed on the player screen. They aim to keep a high frame rate (30 to 60 Frames per seconds) by scheduling the tasks required to complete such computations.
 
-Each task represents a functionality written by a given team at a given moment in the lifetime of the game engine. These tasks have precedences constraints that must be respected to ensure execution's correctness which leads their organisation as a directed acyclic graph (DAG) and have various execution time.
+Each task represents a functionality written by a given team at a given moment in the lifetime of the game engine. These tasks have precedence constraints that must be respected to ensure the execution's correctness which is organised as a directed acyclic graph (DAG) and have various execution time.
 Scheduling such tasks is NP-hard and has not been studied in detail in the context of game engine. A better understanding of this problem could lead to optimizations for the benefit of players and developers.
 
-Our code is focus on modeling a modern game engine and answering the following three questions : 
+Our code is focused on modeling a modern game engine and answering the following three questions :
+
 - Can scheduling strategies from the state of the art improve the performance of game Engine
 - Can changes in the scheduling mechanism reduce the gap between schedulers and critical path?
 - Can small changes to the task graph lead to performance improvements
@@ -70,9 +71,9 @@ The software simulates the behavior of game engines using two files :
 - a dependency file describing links between tasks and the structure of the game engine to be simulated
 - a tasks file describing key values for tasks (steps, mean, minimum, maximum, std_dev) for both moderated and intensive use of the engine.
 
-Using these two files, SimSGamE generates randomized tasks composing a frame respecting the statistical repartition and their dependencies (generation may be seeded in order to enable direct comparisons). The simulation also mimic different load evolution (from moderate use 0% to intensive 100%) and evaluates multiple scheduling techniques to estimate frame span and metrics giving informations regarding heuristics efficiencies.
+Using these two files, SimSGamE generates randomized tasks composing a frame respecting the statistical repartition and their dependencies (generation may be seeded in order to enable direct comparisons). The simulation also mimics different load evolution (from moderate use 0% to intensive 100%) and evaluates multiple scheduling techniques to estimate frame span and metrics giving information regarding heuristics efficiencies.
 
[…]
