### 2022-08-09 17e20be
Update paper.md
Add Europar Doi to paper.md
