+++ paper.md
@@ -48,61 +48,61 @@ europar-doi: 10.1007/978-3-031-12597-3_7
 
 # Summary
 
-The video game market is valued at over 100$ billion USD [@gaming] and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
+The video game market is valued at over USD$100 billion [@gaming] and impacts computing both at the hardware and software levels. It produces and sells tens of millions of video games and consoles yearly.
 The variety of devices running games is very large, from personal computers, consoles and smartphones to Cloud gaming servers.
-In order to develop these games more easily and allow to port them on different platforms, developers use a key software component called the game engine.
+In order to develop these games more easily and allow them to be ported to different platforms, developers use a key software component called the game engine.
 
-Game engines are at the heart of the design of modern video games. They handle most of the components of the game, and allow the generation of each frame(image) displayed on the player screen. They aim to keep a high frame rate (30 to 60 Frames per seconds) by scheduling the tasks required to complete such computations.
+Game engines are at the heart of the design of modern video games. They handle most of the components of the game, and allow the generation of each frame (image) displayed on the player screen. They aim to keep a high frame rate (30 to 60 frames per seconds) by scheduling the tasks required to complete such computations.
 
-Each task represents a functionality written by a given team at a given moment in the lifetime of the game engine. These tasks have precedence constraints that must be respected to ensure the execution's correctness which is organised as a directed acyclic graph (DAG) and have various execution time.
-Scheduling such tasks is NP-hard and has not been studied in detail in the context of game engine. A better understanding of this problem could lead to optimizations for the benefit of players and developers.
+Each task represents a functionality written by a given team at a given moment in the lifetime of the game engine. These tasks have precedence constraints that must be respected to ensure the execution's correctness, which is organised as a directed acyclic graph (DAG), and they have varied execution times.
+Scheduling such tasks is NP-hard and has not been studied in detail in the context of game engines. A better understanding of this problem could lead to optimizations for the benefit of players and developers.
 
-Our code is focused on modeling a modern game engine and answering the following three questions :
+Our code is focused on modeling a modern game engine and answering the following three questions:
 
-- Can scheduling strategies from the state of the art improve the performance of game Engine
-- Can changes in the scheduling mechanism reduce the gap between schedulers and critical path?
-- Can small changes to the task graph lead to performance improvements
+- Can state-of-the-art scheduling strategies improve the performance of the game engine?
+- Can changes in the scheduling mechanism reduce the gap between produced schedules and the critical path?
+- Can small changes to the task graph lead to performance improvements?
 
 
 # Statement of need
 
-SimSGamE was created in order to study the scheduling of game engines. Given the lack of studies on this problem, our efforts have been dedicated to finding and adapting algorithms and heuristics proposed in other contexts. We find that there is value in bringing to light new applications and knowledge on existing algorithms.
+SimSGamE was created in order to study the scheduling of game engines. Given the lack of studies on this problem, our efforts have been dedicated to finding and adapting algorithms and heuristics proposed in other contexts. We find that there is value in bringing to light new applications and knowledge about existing algorithms.
 
-The software simulates the behavior of game engines using two files : 
+The software simulates the behavior of game engines using two files: 
 
-- a dependency file describing links between tasks and the structure of the game engine to be simulated
[…]
