### 2022-08-09 #4592 @danielskatz
@baptisteCoye - I've suggested a number of changes to the paper/bib in https://github.com/baptisteCoye/SimSGame/pull/2 - I think this addresses the reference issue originally brought up by @azoitl. Please merge this, or let me know what you disagree with.  I hope then, after we hear from @azoitl and @hwloidl to confirm their satisfaction with the submission, we can proceed to acceptance and publication.
