### 2022-08-11 #4592 @azoitl
> Hi @azoitl and @hwloidl - pinging again: It looks like you both are satisfied with this to proceed to publication based on your checklists. Is this correct?

@danielskatz  yes this is correct. I'm satisfied with this paper and it is ready for publication.
