### 2022-08-12 #4592 @hwloidl
@baptisteCoye thanks for submitting to EuroPar22 and to the special issue in JOSS as well! The extra effort in making this a validated open-source packages is much appreciated!
